pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arenamcp"
version = "0.2.0"
description = "Real-time MCP server bridging MTGA game logs to Claude for conversational game analysis"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "mcp>=1.0.0",
    "watchdog>=3.0.0",
    "requests>=2.28.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "keyboard>=0.13.5",
]

[project.optional-dependencies]
voice = [
    "faster-whisper>=1.0.0",
    "kokoro>=0.3.0",
    "soundfile>=0.12.0",
]
llm = [
    "google-genai>=1.0.0",
    "anthropic>=0.18.0",
    "ollama>=0.1.0",
]
full = [
    "arenamcp[voice,llm]",
]
dev = [
    "pytest>=7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/arenamcp"]
